# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Solana War Room API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tokens/info:
    post:
      operationId: getTokenInfo
      tags: [tokens]
      summary: Get token info by mint address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenInfoRequest"
      responses:
        "200":
          description: Token info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenInfo"
  /trades/buy:
    post:
      operationId: executeBuy
      tags: [trades]
      summary: Execute buy on selected accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyRequest"
      responses:
        "200":
          description: Trade results per account
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeResult"
  /trades/sell:
    post:
      operationId: executeSell
      tags: [trades]
      summary: Execute sell on selected accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellRequest"
      responses:
        "200":
          description: Trade results per account
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeResult"
  /trades/positions:
    get:
      operationId: getPositions
      tags: [trades]
      summary: Get P&L positions for accounts on a mint
      parameters:
        - name: mintAddress
          in: query
          required: true
          schema:
            type: string
        - name: accountIds
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated account IDs
      responses:
        "200":
          description: P&L positions per account
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountPosition"
  /volume/start:
    post:
      operationId: startVolumeJob
      tags: [volume]
      summary: Start a volume generation job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolumeJobRequest"
      responses:
        "201":
          description: Volume job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolumeJob"
...
  /volume/jobs/{jobId}/stop:
    post:
      operationId: stopVolumeJob
      tags: [volume]
      summary: Stop a running volume job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolumeJob"
...
components:
  schemas:
...
    TokenInfo:
      type: object
      properties:
        mintAddress:
          type: string
        name:
          type: string
        symbol:
          type: string
        decimals:
          type: integer
        price:
          type: number
          nullable: true
        marketCap:
          type: number
          nullable: true
        liquidity:
          type: number
          nullable: true
        volume24h:
          type: number
          nullable: true
        dex:
          type: string
          description: "pump_fun | raydium | orca | jupiter"
        graduated:
          type: boolean
          nullable: true
        logoUri:
          type: string
          nullable: true
      required: [mintAddress, name, symbol, decimals, dex]
    TokenInfoRequest:
      type: object
      properties:
        mintAddress:
          type: string
      required: [mintAddress]
    BuyRequest:
      type: object
      properties:
        mintAddress:
          type: string
        accountIds:
          type: array
          items:
            type: integer
        amountSol:
          type: number
          description: Fixed SOL amount per account (use when not using random range)
        minAmountSol:
          type: number
          description: Min SOL amount for random range mode
        maxAmountSol:
          type: number
          description: Max SOL amount for random range mode
        slippageBps:
          type: integer
          default: 500
          description: Slippage in basis points
        useJito:
          type: boolean
          default: true
        jitoTipLamports:
          type: integer
          default: 10000
        password:
          type: string
          description: Password to decrypt private keys
        delayMs:
          type: integer
          default: 0
          description: Delay between each account transaction in ms
      required: [mintAddress, accountIds, password]
    AccountPosition:
      type: object
      properties:
        accountId:
          type: integer
        accountName:
          type: string
        publicKey:
          type: string
        totalSolIn:
          type: number
        buyCount:
          type: integer
        tokenBalance:
          type: number
        currentValueSol:
          type: number
        estimatedFees:
          type: number
        pnlSol:
          type: number
        pnlPct:
          type: number
      required: [accountId, accountName, publicKey, totalSolIn, buyCount, tokenBalance, currentValueSol, estimatedFees, pnlSol, pnlPct]
    SellRequest:
      type: object
      properties:
        mintAddress:
          type: string
        accountIds:
          type: array
          items:
            type: integer
        percentToSell:
          type: integer
          default: 100
          description: Percentage of token balance to sell (1-100)
        slippageBps:
          type: integer
          default: 500
        useJito:
          type: boolean
          default: true
        jitoTipLamports:
          type: integer
          default: 10000
        password:
          type: string
        delayMs:
          type: integer
          default: 0
      required: [mintAddress, accountIds, percentToSell, password]
    TradeResult:
      type: object
      properties:
        accountId:
          type: integer
        accountName:
          type: string
        publicKey:
          type: string
        status:
          type: string
          enum: [success, failed, pending]
        txSignature:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        amountIn:
          type: number
          nullable: true
        amountOut:
          type: number
          nullable: true
      required: [accountId, accountName, publicKey, status]
    VolumeJobRequest:
      type: object
      properties:
        mintAddress:
          type: string
        accountIds:
          type: array
          items:
            type: integer
        minAmountSol:
          type: number
          description: Min SOL per trade
        maxAmountSol:
          type: number
          description: Max SOL per trade
        minDelayMs:
          type: integer
          description: Min delay between trades in ms
        maxDelayMs:
          type: integer
          description: Max delay between trades in ms
        totalDurationMinutes:
          type: integer
          description: How long to run the job in minutes
        slippageBps:
          type: integer
          default: 500
        useJito:
          type: boolean
          default: true
        jitoTipLamports:
          type: integer
          default: 10000
        password:
          type: string
        pattern:
          type: string
          enum: [random, wash, ladder]
          default: random
          description: "random=random buys/sells between accounts, wash=buy then sell same account, ladder=incremental buys"
      required: [mintAddress, accountIds, minAmountSol, maxAmountSol, minDelayMs, maxDelayMs, totalDurationMinutes, password]
    VolumeJob:
      type: object
      properties:
        id:
          type: integer
        mintAddress:
          type: string
        tokenSymbol:
          type: string
          nullable: true
        status:
          type: string
          enum: [running, stopped, completed, failed]
        pattern:
          type: string
        totalTrades:
          type: integer
        successfulTrades:
          type: integer
        failedTrades:
          type: integer
        totalVolumeSol:
          type: number
        startedAt:
          type: string
          format: date-time
        stoppedAt:
          type: string
          format: date-time
          nullable: true
        endsAt:
          type: string
          format: date-time
          nullable: true
      required: [id, mintAddress, status, pattern, totalTrades, successfulTrades, failedTrades, totalVolumeSol, startedAt]
...